Android: Pass raw pointer of QAndroidScreenCapture as ID
Before this commit, QAndroidScreenCapture implementation used static
map of QAndroidScreenCapture objects. Using a raw pointer as ID allows
to get rid of the static map (and no mutex is needed).

Task-number: QTBUG-112855
Pick-to: 6.8
(cherry picked from commit 2dabdc7dca9c6a0566b3dd35cb3c8903516ab6dc)

+++ src/android/jar/src/org/qtproject/qt/android/multimedia/QtScreenGrabber.java
@@ -73,7 +73,7 @@ class QtScreenGrabber {
         }
     }
 
-    public boolean startScreenCaptureService(int resultCode, int id, int width, int height, Intent data) {
+    public boolean startScreenCaptureService(int resultCode, long id, int width, int height, Intent data) {
         try {
             Intent serviceIntent = new Intent(m_activity, QtScreenCaptureService.class);
             serviceIntent.putExtra(RESULT_CODE, resultCode);
